chore: 🔧 adjusting the directory structure
