+++ packages/x6-common/src/index.ts
@@ -1,6 +1,5 @@
 import './polyfill'
 
-export * from './version'
 export * from './array'
 export * from './object'
 export * from './string'
