--- packages/x6-core/__tests__/version.test.ts
@@ -1,9 +1,0 @@
-import { version } from '../src/global'
-
-describe('version', () => {
-  it('should match the `version` field of package.json', () => {
-    // eslint-disable-next-line
-    const expected = require('../package.json').version
-    expect(version).toBe(expected)
-  })
-})
