+++ packages/x6-core/src/common/config.ts
