+++ packages/x6-core/src/common/index.ts
@@ -1,1 +1,3 @@
 export * from './basecoat'
+export * from './disposable'
+export * from './config'
