--- packages/x6-core/src/global/index.ts
@@ -1,2 +1,0 @@
-export * from './config'
-export * from './version'
