--- packages/x6-core/src/global/version.ts
@@ -1,7 +1,0 @@
-/* eslint-disable */
-
-/**
- * Auto generated version file, do not modify it!
- */
-const version = '1.0.1'
-export { version }
