+++ packages/x6-core/src/index.ts
@@ -1,5 +1,4 @@
 export * from './common'
-export * from './global'
 export * from './model'
 export * from './view'
 export * from './util'
