+++ packages/x6-core/src/registry/highlighter/stroke.ts
@@ -1,9 +1,9 @@
 import { ObjectExt, Dom, Vector } from '@antv/x6-common'
 import { Attr } from '../attr'
-import { Config } from '../../global'
+import { Config } from '../../common'
 import { EdgeView } from '../../view'
 import { Highlighter } from './index'
-import { transformRectangle, bbox } from '../../util'
+import { Util } from '../../util'
 
 export interface StrokeHighlighterOptions {
   padding?: number
@@ -41,7 +41,7 @@ export const stroke: Highlighter.Definition<StrokeHighlighterOptions> = {
     } catch (error) {
       // Failed to get path data from magnet element.
       // Draw a rectangle around the entire cell view instead.
-      magnetBBox = bbox(magnetVel.node, true)
+      magnetBBox = Util.bbox(magnetVel.node, true)
       pathData = Dom.rectToPathData({ ...options, ...magnetBBox })
     }
 
@@ -67,13 +67,13 @@ export const stroke: Highlighter.Definition<StrokeHighlighterOptions> = {
       const padding = options.padding
       if (padding) {
         if (magnetBBox == null) {
-          magnetBBox = bbox(magnetVel.node, true)
+          magnetBBox = Util.bbox(magnetVel.node, true)
         }
 
         const cx = magnetBBox.x + magnetBBox.width / 2
         const cy = magnetBBox.y + magnetBBox.height / 2
 
-        magnetBBox = transformRectangle(magnetBBox, highlightMatrix)
+        magnetBBox = Util.transformRectangle(magnetBBox, highlightMatrix)
 
         const width = Math.max(magnetBBox.width, 1)
         const height = Math.max(magnetBBox.height, 1)
